+++ apps/api/src/app/portfolio/calculator/portfolio-calculator.factory.ts
@@ -1,10 +1,13 @@
-import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interface';
 import { CurrentRateService } from '@ghostfolio/api/app/portfolio/current-rate.service';
 import { RedisCacheService } from '@ghostfolio/api/app/redis-cache/redis-cache.service';
 import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
 import { PortfolioSnapshotService } from '@ghostfolio/api/services/queues/portfolio-snapshot/portfolio-snapshot.service';
-import { Filter, HistoricalDataItem } from '@ghostfolio/common/interfaces';
+import {
+  ActivityResponse,
+  Filter,
+  HistoricalDataItem
+} from '@ghostfolio/common/interfaces';
 import { PerformanceCalculationType } from '@ghostfolio/common/types/performance-calculation-type.type';
 
 import { Injectable } from '@nestjs/common';
@@ -34,7 +37,7 @@ export class PortfolioCalculatorFactory {
     userId
   }: {
     accountBalanceItems?: HistoricalDataItem[];
-    activities: Activity[];
+    activities: ActivityResponse[];
     calculationType: PerformanceCalculationType;
     currency: string;
     filters?: Filter[];