+++ apps/client/src/app/pages/portfolio/activities/create-or-update-activity-dialog/interfaces/interfaces.ts
@@ -1,11 +1,10 @@
-import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interface';
-import { User } from '@ghostfolio/common/interfaces';
+import { ActivityResponse, User } from '@ghostfolio/common/interfaces';
 
 import { Account } from '@prisma/client';
 
 export interface CreateOrUpdateActivityDialogParams {
   accountId: string;
   accounts: Account[];
-  activity: Activity;
+  activity: ActivityResponse;
   user: User;
 }