+++ apps/client/src/app/services/import-activities.service.ts
@@ -2,8 +2,8 @@ import { CreateTagDto } from '@ghostfolio/api/app/endpoints/tags/create-tag.dto'
 import { CreateAccountWithBalancesDto } from '@ghostfolio/api/app/import/create-account-with-balances.dto';
 import { CreateAssetProfileWithMarketDataDto } from '@ghostfolio/api/app/import/create-asset-profile-with-market-data.dto';
 import { CreateOrderDto } from '@ghostfolio/api/app/order/create-order.dto';
-import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interface';
 import { parseDate as parseDateHelper } from '@ghostfolio/common/helper';
+import { ActivityResponse } from '@ghostfolio/common/interfaces';
 
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
@@ -44,7 +44,7 @@ export class ImportActivitiesService {
     isDryRun?: boolean;
     userAccounts: Account[];
   }): Promise<{
-    activities: Activity[];
+    activities: ActivityResponse[];
     assetProfiles: CreateAssetProfileWithMarketDataDto[];
   }> {
     const content = csvToJson(fileContent, {
@@ -128,7 +128,7 @@ export class ImportActivitiesService {
     isDryRun?: boolean;
     tags?: CreateTagDto[];
   }): Promise<{
-    activities: Activity[];
+    activities: ActivityResponse[];
   }> {
     return new Promise((resolve, reject) => {
       this.postImport(
@@ -161,11 +161,11 @@ export class ImportActivitiesService {
     tags
   }: {
     accounts?: CreateAccountWithBalancesDto[];
-    activities: Activity[];
+    activities: ActivityResponse[];
     assetProfiles?: CreateAssetProfileWithMarketDataDto[];
     tags?: CreateTagDto[];
   }): Promise<{
-    activities: Activity[];
+    activities: ActivityResponse[];
   }> {
     const importData: CreateOrderDto[] = [];