+++ libs/common/src/lib/interfaces/index.ts
@@ -37,7 +37,11 @@ import type { Product } from './product';
 import type { AccessTokenResponse } from './responses/access-token-response.interface';
 import type { AccountBalancesResponse } from './responses/account-balances-response.interface';
 import type { AccountsResponse } from './responses/accounts-response.interface';
-import type { ActivityResponse } from './responses/activity-response.interface';
+import type {
+  ActivitiesResponse,
+  ActivityErrorResponse,
+  ActivityResponse
+} from './responses/activity-response.interface';
 import type { AiPromptResponse } from './responses/ai-prompt-response.interface';
 import type { ApiKeyResponse } from './responses/api-key-response.interface';
 import type { AssetResponse } from './responses/asset-response.interface';
@@ -84,6 +88,8 @@ export {
   AccountBalance,
   AccountBalancesResponse,
   AccountsResponse,
+  ActivitiesResponse,
+  ActivityErrorResponse,
   ActivityResponse,
   AdminData,
   AdminJobs,